requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
openai==1.54.4
python-dotenv==1.0.0
sqlparse==0.4.4
httpx==0.25.2